[1월 2주차] jqGrid Up/Down 기능, 페이징 넘버링 추가

+++ src/main/java/study/board/service/impl/BoardServiceImpl.java
@@ -60,7 +60,7 @@ public class BoardServiceImpl implements BoardService{
 		
 		// 선택된 행 데이터 삭제
 		int result = 0;
-		if (boardId != null) {
+		if (boardId != null && !boardId.isEmpty()) {
 			result = boardDAO.deleteBoardList(boardId);
 		}
